+++ packages/puzzle/examples/overlays/app/routes.js
@@ -1,11 +1,11 @@
-import { lazy } from '@magic-spells/puzzle';
+import HomeView from './views/Home.pzl';
 import DefaultLayout from './layouts/Default.pzl';
 
 export default [
   {
     path: '/',
     name: 'home',
-		view: lazy(() => import('./views/Home.pzl')),
+    view: HomeView,
     layout: DefaultLayout,
     meta: {
       title: 'Overlays — Puzzle <Portal>',